+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [5.0.0-rc.1] - 2025-11-14
 
 ### Added
 
@@ -284,7 +284,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
-[Unreleased]: https://https://github.com/internetguru/laravel-feedback/compare/staging...dev
+[5.0.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v4.2.0
 [4.2.0]: https://https://github.com/internetguru/laravel-feedback/compare/v4.1.0...v4.2.0
 [4.2.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v4.1.0
 [4.1.0]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.2...v4.1.0